Add cargo deny
Adds an initial config file and the tool to the `[tools]` section
of rust-toolchain.toml so it is installed with the rust tool cache
action in the CI workflow. Adds relevant exceptions with notes.

Adds the `cargo make deny` task to the makefile to run the command
in this repo.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>